# openapi.yaml
openapi: 3.1.0
info:
  title: AgentMail API
  version: 1.0.0
  description: >
    API for giving AI agents their own email inboxes — send, receive, and
    manage threaded conversations programmatically.
  x-organization: agentmail-demo
servers:
  - url: https://agentmail-demo.vercel.app
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key (sk_live_...)
  schemas:
...
    Message:
      type: object
      required: [id, thread_id, inbox_email_address, direction, message_id_header, from, to, cc, bcc, headers, status, created_at]
      properties:
        id:
          type: string
          description: RFC 5322 Message-ID (normalized, without angle brackets)
          example: 1234567890.abcdef.12345@yourdomain.com
        thread_id:
          type: string
          example: thread_abc123
        inbox_email_address:
          type: string
          format: email
          example: alice@yourdomain.com
        direction:
          type: string
          enum: [inbound, outbound]
        message_id_header:
          type: string
        in_reply_to:
          type: [string, "null"]
        from:
          type: string
          format: email
        to:
          type: array
          items:
            type: string
            format: email
        cc:
          type: array
          items:
            type: string
            format: email
        bcc:
          type: array
          items:
            type: string
            format: email
        subject:
          type: [string, "null"]
        body_text:
          type: [string, "null"]
        body_html:
          type: [string, "null"]
        headers:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [sent, received, delivered, bounced]
        created_at:
          type: string
          format: date-time
    Thread:
      type: object
      required: [id, inbox_email_address, root_message_id_header, created_at, messages]
      properties:
        id:
          type: string
          example: thread_abc123
        inbox_email_address:
          type: string
          format: email
          example: alice@yourdomain.com
        subject:
          type: [string, "null"]
        root_message_id_header:
          type: string
        last_message_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    ThreadWithSimilarity:
      allOf:
        - $ref: "#/components/schemas/Thread"
        - type: object
          required: [similarity]
          properties:
            similarity:
              type: number
              format: float
              minimum: 0
              maximum: 1
    WebhookEndpoint:
      type: object
      required: [id, url, is_enabled, created_at]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        description:
          type: [string, "null"]
        subscribed_events:
          type: [array, "null"]
          items:
            type: string
          example: [message.received, thread.created]
        is_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    WebhookEndpointWithSecret:
      allOf:
        - $ref: "#/components/schemas/WebhookEndpoint"
        - type: object
          required: [signing_secret]
          properties:
            signing_secret:
              type: string
              description: Shown only on creation and after rotation.
...
    ThreadList:
      type: object
      required: [object, data, has_more]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Thread"
        has_more:
          type: boolean
        next_cursor:
          type: [string, "null"]
    ThreadSearchList:
      type: object
      required: [object, data, has_more]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/ThreadWithSimilarity"
        has_more:
          type: boolean
        next_cursor:
          type: [string, "null"]
...
    WebhookEndpointList:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEndpoint"
    DeletedResource:
      type: object
      required: [deleted, id]
      properties:
        deleted:
          type: boolean
          enum: [true]
        id:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
paths:
...
  /v1/threads:
    get:
      operationId: listThreads
      summary: List threads
      description: Returns threads for a given inbox, ordered by most recent activity.
      tags: [Threads]
      parameters:
        - name: inbox_email_address
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: starting_after
          in: query
          description: Cursor for pagination (thread ID of the last item).
          schema:
            type: string
        - name: participant
          in: query
          description: Filter threads by an email address that appears as sender or recipient.
          schema:
            type: string
            format: email
      responses:
        "200":
          description: List of threads (each with nested messages)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadList"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Inbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/threads/{id}:
    get:
      operationId: getThread
      summary: Get thread
      description: Retrieve a thread with all its messages.
      tags: [Threads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thread with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/threads/search:
    post:
      operationId: searchThreads
      summary: Semantic search threads
      description: >
        Performs vector similarity search across thread messages using pgvector
        embeddings (Xenova/all-MiniLM-L6-v2). Requires embeddings to be enabled.
      tags: [Threads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Natural-language search query.
                inbox_email_address:
                  type: string
                  format: email
                  description: Optional — restrict search to a single inbox.
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 10
      responses:
        "200":
          description: Matching threads ordered by similarity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadSearchList"
        "400":
          description: Validation error or embeddings not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Inbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/webhook_endpoints:
    post:
      operationId: createWebhookEndpoint
      summary: Create webhook endpoint
      description: >
        Register a URL to receive event deliveries. The signing_secret is returned
        only on creation — store it securely.
      tags: [Webhook Endpoints]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: Must be HTTPS.
                description:
                  type: string
                subscribed_events:
                  type: array
                  items:
                    type: string
                  description: >
                    Event types to receive. Defaults to all events when omitted.
                  example:
                    - message.received
                    - thread.created
      responses:
        "201":
          description: Webhook endpoint created (includes signing_secret)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointWithSecret"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      operationId: listWebhookEndpoints
      summary: List webhook endpoints
      tags: [Webhook Endpoints]
      responses:
        "200":
          description: List of webhook endpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointList"
  /v1/webhook_endpoints/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getWebhookEndpoint
      summary: Get webhook endpoint
      description: Returns the endpoint. signing_secret is NOT included.
      tags: [Webhook Endpoints]
      responses:
        "200":
          description: Webhook endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      operationId: updateWebhookEndpoint
      summary: Update webhook endpoint
      tags: [Webhook Endpoints]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                description:
                  type: [string, "null"]
                subscribed_events:
                  type: array
                  items:
                    type: string
                is_enabled:
                  type: boolean
      responses:
        "200":
          description: Updated webhook endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      operationId: deleteWebhookEndpoint
      summary: Delete webhook endpoint
      tags: [Webhook Endpoints]
      responses:
        "200":
          description: Deletion confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/webhook_endpoints/{id}/rotate_secret:
    post:
      operationId: rotateWebhookSecret
      summary: Rotate signing secret
      description: Generates a new signing secret. Returns the new value once — store it immediately.
      tags: [Webhook Endpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: New signing secret
          content:
            application/json:
              schema:
                type: object
                required: [id, signing_secret]
                properties:
                  id:
                    type: string
                    format: uuid
                  signing_secret:
                    type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...